Very large public API typing rework to use more portable API typing.
This may have an effect on existing code.  Concretely, some typecast
or signed/unsigned warnings may appear for existing calling code.

With this change, the public API should be more or less clean for
porting to platforms with a 16-bit 'int' type.

+++ api-testcases/test-push-global.c
@@ -6,13 +6,12 @@ duk_is_function(global.escape)=1
 
 void test(duk_context *ctx) {
 	duk_push_global_object(ctx);
-	printf("type=%d\n", duk_get_type(ctx, -1));
+	printf("type=%d\n", (int) duk_get_type(ctx, -1));
 	printf("ToString=%s\n", duk_to_string(ctx, -1));
 	duk_pop(ctx);
 
 	duk_push_global_object(ctx);
 	duk_get_prop_string(ctx, -1, "escape");
-	printf("duk_is_function(global.escape)=%d\n", duk_is_function(ctx, -1));
+	printf("duk_is_function(global.escape)=%d\n", (int) duk_is_function(ctx, -1));
 	duk_pop(ctx);
 }
-
